Gebruik RGBZ Eigenschappen om extra informatie aan de zaak te koppelen
