+++ src/openapi.yaml
@@ -31,46 +31,6 @@ paths:
         required: true
         schema:
           type: integer
-  /domeindata:
-    post:
-      operationId: domeindata_create
-      description: Registreer DOMEINDATA bij een zaak.
-      responses:
-        '201':
-          description: ''
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/DomeinData'
-      tags:
-        - domeindata
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/DomeinData'
-        required: true
-    parameters: []
-  '/domeindata/{id}':
-    get:
-      operationId: domeindata_read
-      description: Geef de details van DOMEINDATA voor een ZAAK.
-      responses:
-        '200':
-          description: ''
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/DomeinData'
-      tags:
-        - domeindata
-    parameters:
-      - name: id
-        in: path
-        description: A unique integer value identifying this domeindatareferentie.
-        required: true
-        schema:
-          type: integer
   /klantcontacten:
     post:
       operationId: klantcontact_create
@@ -193,6 +153,46 @@ paths:
       - name: id
         in: path
         description: A unique integer value identifying this status.
+        required: true
+        schema:
+          type: integer
+  /zaakeigenschappen:
+    post:
+      operationId: zaakeigenschap_create
+      description: Registreer een eigenschap van een zaak.
+      responses:
+        '201':
+          description: ''
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/ZaakEigenschap'
+      tags:
+        - zaakeigenschappen
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/ZaakEigenschap'
+        required: true
+    parameters: []
+  '/zaakeigenschappen/{id}':
+    get:
+      operationId: zaakeigenschap_read
+      description: Geef de details van ZaakEigenschap voor een ZAAK.
+      responses:
+        '200':
+          description: ''
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/ZaakEigenschap'
+      tags:
+        - zaakeigenschappen
+    parameters:
+      - name: id
+        in: path
+        description: A unique integer value identifying this zaakeigenschap.
         required: true
         schema:
           type: integer
@@ -327,28 +327,6 @@ components:
           type: string
           maxLength: 50
           minLength: 1
-    DomeinData:
-      required:
-        - zaak
-        - domeinData
-      type: object
-      properties:
-        url:
-          title: Url
-          type: string
-          format: uri
-          readOnly: true
-        zaak:
-          title: Zaak
-          type: string
-          format: uri
-        domeinData:
-          title: Domein data
-          description: URL naar de domein data resource
-          type: string
-          format: uri
-          maxLength: 200
-          minLength: 1
     KlantContact:
       required:
         - zaak
@@ -466,6 +444,33 @@ components:
             status van een zaak.
           type: string
           maxLength: 1000
+    ZaakEigenschap:
+      required:
+        - zaak
+        - eigenschap
+        - waarde
+      type: object
+      properties:
+        url:
+          title: Url
+          type: string
+          format: uri
+          readOnly: true
+        zaak:
+          title: Zaak
+          type: string
+          format: uri
+        eigenschap:
+          title: Eigenschap
+          description: URL naar de eigenschap in het ZTC
+          type: string
+          format: uri
+          maxLength: 200
+          minLength: 1
+        waarde:
+          title: Waarde
+          type: string
+          minLength: 1
     ZaakObject:
       required:
         - zaak
