+++ src/zrc/api/urls.py
@@ -4,15 +4,15 @@ from rest_framework.routers import DefaultRouter
 
 from .schema import schema_view
 from .viewsets import (
-    BetrokkeneViewSet, DomeinDataViewSet, KlantContactViewSet, RolViewSet,
-    StatusViewSet, ZaakObjectViewSet, ZaakViewSet
+    BetrokkeneViewSet, KlantContactViewSet, RolViewSet, StatusViewSet,
+    ZaakEigenschapViewSet, ZaakObjectViewSet, ZaakViewSet
 )
 
 router = DefaultRouter(trailing_slash=False)
 router.register('zaken', ZaakViewSet)
 router.register('statussen', StatusViewSet)
 router.register('zaakobjecten', ZaakObjectViewSet)
-router.register('domeindata', DomeinDataViewSet)
+router.register('zaakeigenschappen', ZaakEigenschapViewSet)
 router.register('klantcontacten', KlantContactViewSet)
 router.register('betrokkenen', BetrokkeneViewSet)
 router.register('rollen', RolViewSet)
