+++ src/zrc/templates/master.html
@@ -26,7 +26,5 @@
                 to get some actual content.
             {% endblock %}
         </main>
-
-        <script src="{% static 'js/zrc.js' %}"></script>
     </body>
 </html>
