+++ src/zrc/tests/test_userstory_39.py
@@ -3,52 +3,25 @@ Test the flow described in https://github.com/VNG-Realisatie/gemma-zaken/issues/
 """
 from datetime import date
 
-from dateutil import parser
 from rest_framework import status
 from rest_framework.test import APITestCase
 from zds_schema.tests import get_operation_url
 
-from zrc.datamodel.models import (
-    DomeinData, KlantContact, Rol, Status, Zaak, ZaakObject
-)
+from zrc.datamodel.models import KlantContact, Rol, Status, Zaak, ZaakObject
 from zrc.datamodel.tests.factories import (
     OrganisatorischeEenheidFactory, ZaakFactory
 )
 
-from .utils import isodatetime, utcdatetime
+from .utils import isodatetime
 
 ZAAKTYPE = 'https://example.com/ztc/api/v1/catalogus/1/zaaktypen/1'
 STATUS_TYPE = 'https://example.com/ztc/api/v1/catalogus/1/zaaktypen/1/statustypen/1'
 STATUS_TYPE_OVERLAST_GECONSTATEERD = 'https://example.com/ztc/api/v1/catalogus/1/zaaktypen/1/statustypen/2'
 OBJECT_MET_ADRES = 'https://example.com/orc/api/v1/objecten/1'
-DOMEIN_DATA = 'https://example.com/domeindata/api/v1/data/1'
 FOTO = 'https://example.com/drc/api/v1/enkelvoudiginformatieobjecten/1'
 # file:///home/bbt/Downloads/2a.aansluitspecificatieskennisgevingen-gegevenswoordenboek-entiteitenv1.0.6.pdf
 # Stadsdeel is een WijkObject in het RSGB
 STADSDEEL = 'https://example.com/rsgb/api/v1/wijkobjecten/1'
-
-TEST_DATA = {
-    "id": 9966,
-    "last_status": "o",
-    "adres": "Oosterdok 51, 1011 Amsterdam, Netherlands",
-    "datetime": "2018-05-28T09:05:08.732587+02:00",
-    "text": "test",
-    "waternet_soort_boot": "Nee",
-    "waternet_rederij": "Onbekend",
-    "waternet_naam_boot": "",
-    "datetime_overlast": "2018-05-28T08:35:11+02:00",
-    "email": "",
-    "phone_number": "",
-    "source": "Telefoon 14020",
-    "text_extra": "",
-    "image": None,
-    "main_category": "",
-    "sub_category": "Geluid",
-    "ml_cat": "melding openbare ruimte",
-    "stadsdeel": "Centrum",
-    "coordinates": "POINT (4.910649523925713 52.37240093589432)",
-    "verantwoordelijk": "Waternet"
-}
 
 
 class US39TestCase(APITestCase):
@@ -156,31 +129,6 @@ class US39TestCase(APITestCase):
             }
         )
 
-    def test_zet_domeindata(self):
-        url = get_operation_url('domeindata_create')
-        zaak = ZaakFactory.create()
-        zaak_url = get_operation_url('zaak_read', id=zaak.id)
-        data = {
-            'zaak': zaak_url,
-            'domeinData': DOMEIN_DATA,
-        }
-
-        response = self.client.post(url, data)
-
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        response_data = response.json()
-        domeindata = DomeinData.objects.get()
-        self.assertEqual(domeindata.zaak, zaak)
-        detail_url = get_operation_url('domeindata_read', id=domeindata.id)
-        self.assertEqual(
-            response_data,
-            {
-                'url': f"http://testserver{detail_url}",
-                'zaak': f"http://testserver{zaak_url}",
-                'domeinData': DOMEIN_DATA,
-            }
-        )
-
     def test_create_klantcontact(self):
         url = get_operation_url('klantcontact_create')
         zaak = ZaakFactory.create()
@@ -272,114 +220,4 @@ class US39TestCase(APITestCase):
                 'rolomschrijvingGeneriek': 'Behandelaar',
                 'roltoelichting': 'Baggeren van gracht',
             }
-        )
-
-
-class Application:
-
-    def __init__(self, client, data: dict):
-        self.client = client
-
-        self.data = data
-        self.references = {}
-
-    def store_notification(self):
-        # registreer zaak & zet statussen
-        self.registreer_zaak()
-        self.zet_statussen()
-        self.registreer_domein_data()
-        self.registreer_klantcontact()
-
-    @property
-    def domein_data_url(self):
-        return self.references['domein_data_url']
-
-    def registreer_zaak(self):
-        zaak_create_url = get_operation_url('zaak_create')
-
-        created = parser.parse(self.data['datetime'])
-        intern_id = self.data['id']
-
-        response = self.client.post(zaak_create_url, {
-            'zaaktype': ZAAKTYPE,
-            'zaakidentificatie': f'AMS{intern_id}',
-            'registratiedatum': created.strftime('%Y-%m-%d'),
-            'toelichting': self.data['text'],
-            'zaakgeometrie': self.data['coordinates'],
-        })
-        self.references['zaak_url'] = response.json()['url']
-
-    def zet_statussen(self):
-        status_create_url = get_operation_url('status_create')
-
-        created = parser.parse(self.data['datetime'])
-
-        self.client.post(status_create_url, {
-            'zaak': self.references['zaak_url'],
-            'statusType': STATUS_TYPE,
-            'datumStatusGezet': created.isoformat(),
-        })
-
-        self.client.post(status_create_url, {
-            'zaak': self.references['zaak_url'],
-            'statusType': STATUS_TYPE_OVERLAST_GECONSTATEERD,
-            'datumStatusGezet': parser.parse(self.data['datetime_overlast']).isoformat(),
-        })
-
-    def registreer_domein_data(self):
-        url = get_operation_url('domeindata_create')
-        response = self.client.post(url, {
-            'zaak': self.references['zaak_url'],
-            'domeinData': DOMEIN_DATA,
-        })
-        self.references['domein_data_url'] = response.json()['url']
-
-    def registreer_klantcontact(self):
-        url = get_operation_url('klantcontact_create')
-        self.client.post(url, {
-            'zaak': self.references['zaak_url'],
-            'datumtijd': self.data['datetime'],
-            'kanaal': self.data['source'],
-        })
-
-
-class US39IntegrationTestCase(APITestCase):
-    """
-    Simulate a full realistic flow.
-    """
-
-    def test_full_flow(self):
-        app = Application(self.client, TEST_DATA)
-
-        app.store_notification()
-
-        zaak = Zaak.objects.get(zaakidentificatie='AMS9966')
-        self.assertEqual(zaak.toelichting, 'test')
-        self.assertEqual(zaak.zaakgeometrie.x, 4.910649523925713)
-        self.assertEqual(zaak.zaakgeometrie.y, 52.37240093589432)
-
-        self.assertEqual(zaak.status_set.count(), 2)
-
-        last_status = zaak.status_set.order_by('-datum_status_gezet').first()
-        self.assertEqual(last_status.status_type, STATUS_TYPE)
-        self.assertEqual(
-            last_status.datum_status_gezet,
-            utcdatetime(2018, 5, 28, 7, 5, 8, 732587),
-        )
-
-        first_status = zaak.status_set.order_by('datum_status_gezet').first()
-        self.assertEqual(first_status.status_type, STATUS_TYPE_OVERLAST_GECONSTATEERD)
-        self.assertEqual(
-            first_status.datum_status_gezet,
-            utcdatetime(2018, 5, 28, 6, 35, 11)
-        )
-
-        domein_data = self.client.get(app.domein_data_url).json()['domeinData']
-        self.assertEqual(domein_data, DOMEIN_DATA)
-
-        klantcontact = zaak.klantcontact_set.get()
-        self.assertEqual(klantcontact.kanaal, 'Telefoon 14020')
-        self.assertEqual(
-            klantcontact.datumtijd,
-            utcdatetime(2018, 5, 28, 7, 5, 8, 732587),
         )
